PACT-442 | EPR | Creation of session token after login
Changes after code review
