+++ test/controllers/EditSetControllerSpec.scala
@@ -27,7 +27,7 @@ import play.api.mvc.{ AnyContentAsEmpty, DefaultActionBuilder, DefaultMessagesAc
 import play.api.test._
 import play.api.test.Helpers._
 import uk.gov.nationalarchives.omega.editorial.controllers.{ EditSetController, LoginController }
-import uk.gov.nationalarchives.omega.editorial.models.dao.SessionDAO
+import uk.gov.nationalarchives.omega.editorial.models.session.Session
 
 /** Add your spec here.
   * You can mock out a whole application including requests, plugins etc.
@@ -36,8 +36,8 @@ import uk.gov.nationalarchives.omega.editorial.models.dao.SessionDAO
   */
 class EditSetControllerSpec extends PlaySpec with GuiceOneAppPerTest with Injecting {
 
-  val validSessionToken = SessionDAO.generateToken("1234")
-  val invalidSessionToken = SessionDAO.generateToken("invalid-user")
+  val validSessionToken = Session.generateToken("1234")
+  val invalidSessionToken = Session.generateToken("invalid-user")
 
   "EditSetController GET /edit-set/{id}" should {
